+++ src/webview/client/clothing/app.js
@@ -27,7 +27,12 @@ const app = new Vue({
     },
     methods: {
         setURL(url) {
-            this.url = `http://${url}:9111`;
+            if (url.includes('assets/webserver')) {
+                this.url = url;
+                return;
+            }
+
+            this.url = `http://${url}`;
         },
         incrementComponent(index, dataName) {
             const newComponents = [...this.components];