+++ src/webview/client/creator/app.js
@@ -88,7 +88,12 @@ const app = new Vue({
     },
     methods: {
         setURL(url) {
-            this.url = `http://${url}:9111`;
+            if (url.includes('assets/webserver')) {
+                this.url = url;
+                return;
+            }
+
+            this.url = `http://${url}`;
         },
         incrementIndex() {
             if (this.selection + 1 >= this.navOptions.length) {