+++ src/webview/client/factions/app.js
@@ -23,7 +23,12 @@ const app = new Vue({
     },
     methods: {
         setURL(url) {
-            this.url = `http://${url}:9111`;
+            if (url.includes('assets/webserver')) {
+                this.url = url;
+                return;
+            }
+
+            this.url = `http://${url}`;
         },
         showResponse(response) {
             if (!response) {