+++ src/webview/client/job/app.js
@@ -25,7 +25,12 @@ const app = new Vue({
     },
     methods: {
         setURL(url) {
-            this.url = `http://${url}:9111`;
+            if (url.includes('assets/webserver')) {
+                this.url = url;
+                return;
+            }
+
+            this.url = `http://${url}`;
         },
         setData(jobData) {
             this.image = jobData.image;